Clean up and fix docs
